Modified to 6.03 - did some clean-up as per namcap and modified deps

+++ PKGBUILD
@@ -1,22 +1,24 @@
-# Maintainer: David Wu <xdavidwuph@gmail.com>
+# Maintainer: Vryali <vryali@gmail.com>
+# Contributors: Maintainer: David Wu <xdavidwuph@gmail.com>
 # Contributors: Christopher Pence <chris@pencels.net>
 
 pkgname=xscreensaver-backends
 _srcname=xscreensaver
-pkgver=6.01
+pkgver=6.03
 pkgrel=1
 pkgdesc="Screensavers from XScreenSaver for common frontends"
 arch=('i686' 'x86_64')
 url="http://www.jwz.org/xscreensaver/"
 license=('BSD')
 conflicts=('xscreensaver' 'xscreensaver-hacks' 'mate-screensaver-hacks')
-depends=('libxmu' 'glu' 'libxi')
+#depends=('libxmu' 'glu' 'libxi' 'gdk-pixbuf-xlib')
+depends=('perl' 'gtk2' 'glu' 'gdk-pixbuf-xlib' 'libxt')
 makedepends=('bc' 'intltool' 'libxslt' 'openmotif')
 source=(http://www.jwz.org/xscreensaver/${_srcname}-${pkgver}.tar.gz
 	LICENSE
 	migrate-xscreensaver-config.sh
 	xscreensaver-config.xsl)
-sha1sums=('5e6bf477d14b8a4a07e65ac2fd32b7b7f71e422b'
+sha1sums=('04b272b7b4eadb68e812c682148162bd59d3b9f8'
           '3eedb8b91b13c29df9b1fe5cbb027e1470b802d2'
           '19195ef59f0dbc61c6ed599a968213a8f0a7a5d4'
           'e561e36c7bae61c3f5da65e4fb19a22e698f4584')
@@ -42,11 +44,13 @@ package() {
   cd ${_srcname}-${pkgver}
   make install_prefix="${pkgdir}" install
   install -D -m644 ../LICENSE "${pkgdir}/usr/share/licenses/${pkgname}/LICENSE"
-  rm -r "${pkgdir}/usr/bin" "${pkgdir}/usr/share/locale" "${pkgdir}/usr/share/man/man1" "${pkgdir}/usr/share/X11"
   mkdir -p "${pkgdir}/usr/share/applications/screensavers/"
   cd "${pkgdir}/usr/share/applications/screensavers/"
   sh "${srcdir}/migrate-xscreensaver-config.sh" "${pkgdir}/usr/share/xscreensaver/config/*.xml"
   rm -r "${pkgdir}/usr/share/xscreensaver"
+  rmdir "${pkgdir}/etc"
   # popsquares exists on most frontends
-  rm "${pkgdir}/usr/share/applications/screensavers/popsquares.desktop"
+  if [ -d "${pkgdir}/usr/share/applications/screensavers/popsquares.desktop" ]; then
+    rm "${pkgdir}/usr/share/applications/screensavers/popsquares.desktop"
+  fi
 }
